HRManagement/tools/simulator/requirements.txt
Pillow>=10.0.0
qrcode[pil]>=7.4
requests>=2.31
